Decode codes to plain text (and vice versa)

Translate coded values into meaningful plain text (or reversed).

code(y, keyvalue, verbose = TRUE)

decode(x, ...)

# S3 method for data.frame
decode(x, ...)

# S3 method for default
decode(x, keyvalue, extra_functions = NULL,
  exact = FALSE, ...)

Arguments

y
value to be coded (to be matched againt the value element) in a keyvalue object.

keyvalue
either a name (as character string) of a package internal keyvalue object, or a user defined keyvalue object (see as.keyvalue).

verbose
(only for code) can be set to FALSE to avoid a printed message to the console if an error occur (TRUE as default).

x
object to decode. Either a key vector to be matched againt the key column in keyvalue, or a data.frame (see section decode.data.frame). object.

...
ignored

extra_functions
is a list of functions (or names of functions as character vector) to be applied to the decoded data after decoding (see section "extra_functions" below).

exact
Should x have an exact match from the key? Default is FALSE. When FALSE, x might be transformed to fit the key (punctuation might be removed, upper case changed to lower case or vice versa and strings that are too long might be substringed). (code only accept excact matches.)

Value

For default S3 method: A vector of the same length as x but with all cells decoded (or coded) to plain text (or code) as character.

For S3 method for class 'data.frame': Data.frame x is returned, possibly with some extra columns (names ending in '_Beskrivning'), decoded from columns with names corresponding to attribute standard_var_names for keyvalue objects listed by list_keyvalues().

Vignette

See the vignette for a longer introduction to the package: vignette("decoder")

decode.data.frame

If x is a data.frame, all column names of x are matched to attribute standard_var_names for all keyvalue objects in the package (see list_keyvalues()). If the column name is a standard name used for a coding, the corresponding keyvalue object is used to decode the column and to add an extra column to x with its original name with suffix _Beskrivning. This is done for all identified columns.

extra_functions

The relationship betwen the key and the value in a keyvalue object is either 1:1 or m:1. The mapping is straight forward for 1:1 but with m:1, different applications might require slightly different groupings of the keys. One solution is to have several versions of the keyvalue object. Another (which we prefer) is to use the same keyvalue object (if possible) but to call one or several extra function(s) to further process the result. These functions are either built in package functions that should be called by quoted names or user defined functions that can be called by either quoted or unquoted names (if available in the current environment). Note that the order of the functions could matter since they are called in turn (the output from the first function is passed as input to the second function etc).

Standard functions and how to use them:

To use with sjukvardsomrade:

kungalv2Fyrbodal

The default classification used in sjukvardsomrade is to make Kungalv a region of its own. Use this function if Kungalv should be included in Fyrbodal. See example section below.

kungalv2Storgoteborg

As kungalv2Fyrbodal but classifies Kungalv as a part of Storgoteborg.

real_names

Give the area names with correct Swedish spelling (including spaces). This is not as default due to compability reasons and because names with spaces must be back-ticked when reffered to.

To use with region

short_region_names

Exclude the prefix 'Region' from the region names, hence 'Syd' instead of 'Region Syd' etcetera.

See also

Examples

KON_VALUE <- sample(1:2, 20, replace = TRUE)
(kon <- decode(KON_VALUE, "kon"))

#>  [1] "Man"    "Kvinna" "Kvinna" "Man"    "Man"    "Man"    "Man"    "Man"   
#>  [9] "Kvinna" "Kvinna" "Kvinna" "Man"    "Man"    "Man"    "Man"    "Man"   
#> [17] "Man"    "Man"    "Man"    "Kvinna"

code(kon, "kon")

#>  [1] "1" "2" "2" "1" "1" "1" "1" "1" "2" "2" "2" "1" "1" "1" "1" "1" "1" "1" "1"
#> [20] "2"

# Get a sample of Snomed-codes (in the real world we obviously avoid this step) ...
snomed2 <- sample(decoder:::snomed$key, 30, replace = TRUE)
# ... then decode them:
(snomed3 <- decode(snomed2, "snomed"))

#>  [1] "Misstänkt adenoskvamös cancer"                                      
#>  [2] "Meningiom, fibröst"                                                 
#>  [3] "T-lymfoblastlymfom"                                                 
#>  [4] "Misstänkt adnexcancer"                                              
#>  [5] "Primär malign tumör i lymfoid eller blodbildande vävnad UNS"        
#>  [6] "Subkutant pannikulitiskt T-cellslymfom"                             
#>  [7] "Morbus Bowen"                                                       
#>  [8] "Juvenilt angiofibrom"                                               
#>  [9] "Malign perifer nervskidetumör med rhabdomyoblastisk differentiering"
#> [10] "Misstänkt neurofibrosarkom"                                         
#> [11] "Oatcellscancer, småcellig odifferentierad cancer"                   
#> [12] "Kystadenopapillär cancer"                                           
#> [13] "Tymom, medullärt, malignt"                                          
#> [14] "Medullomyoblastom"                                                  
#> [15] "Misstänkt germinom"                                                 
#> [16] "Mesodermal blandtumör"                                              
#> [17] "T-cells lymfom UNS"                                                 
#> [18] "Akut myeloisk leukemi med multilinjär dysplasi"                     
#> [19] "Huvudcellsadenom"                                                   
#> [20] "Kystadenocarcinom"                                                  
#> [21] "Angioblastiskt meningiom"                                           
#> [22] "Dermoidcysta med malign omvandling"                                 
#> [23] "Misstänkt ganglioneuroblastom"                                      
#> [24] "Teratoid/rhabdoid tumör, atypisk"                                   
#> [25] "Adenocarcinom, intestinal typ"                                      
#> [26] "HD nodulär skleros cellulär fas"                                    
#> [27] "HD nodulär skleros lymfocytic depletion"                            
#> [28] "Refraktär anemi med blastöverskott"                                 
#> [29] "Kronisk myeloproliferativ sjd, UNS"                                 
#> [30] "Fibrosarkom"                                                        

# Health care regions can be defined in more than one way
# By default Kungalv define a region of its own:
set.seed(123456789)
healtcare_areas_west <- sample(unlist(decoder:::sjukvardsomrade), 100, replace = TRUE)
(areas <- decode(healtcare_areas_west, "sjukvardsomrade"))

#> Warning: transformed to match the keyvalue:  Punctuations are removed. Only the first 4 characters are used.

#> Warning: Some codes could not be translated (54 cells)

#>   [1] NA               NA               NA               NA              
#>   [5] NA               NA               "Fyrbodal"       NA              
#>   [9] "Sodra_Alvsborg" "Skaraborg"      "Kungalv"        NA              
#>  [13] NA               "Sodra_Alvsborg" "Skaraborg"      "Sodra_Alvsborg"
#>  [17] "Kungalv"        "Storgoteborg"   NA               NA              
#>  [21] "Kungalv"        NA               NA               "Storgoteborg"  
#>  [25] "Kungalv"        NA               "Sodra_Alvsborg" "Skaraborg"     
#>  [29] NA               NA               "Fyrbodal"       NA              
#>  [33] "Kungalv"        NA               NA               "Skaraborg"     
#>  [37] NA               "Skaraborg"      "Norra_Halland"  "Skaraborg"     
#>  [41] NA               NA               NA               NA              
#>  [45] "Storgoteborg"   NA               NA               "Skaraborg"     
#>  [49] "Fyrbodal"       "Norra_Halland"  NA               NA              
#>  [53] NA               "Sodra_Alvsborg" "Fyrbodal"       "Kungalv"       
#>  [57] "Fyrbodal"       "Skaraborg"      "Skaraborg"      NA              
#>  [61] NA               NA               "Storgoteborg"   "Storgoteborg"  
#>  [65] "Sodra_Alvsborg" "Sodra_Alvsborg" NA               NA              
#>  [69] NA               NA               NA               NA              
#>  [73] "Skaraborg"      NA               NA               "Sodra_Alvsborg"
#>  [77] "Norra_Halland"  NA               NA               NA              
#>  [81] NA               NA               "Skaraborg"      "Storgoteborg"  
#>  [85] NA               NA               NA               NA              
#>  [89] NA               NA               "Fyrbodal"       NA              
#>  [93] "Sodra_Alvsborg" NA               "Fyrbodal"       "Fyrbodal"      
#>  [97] NA               "Skaraborg"      "Fyrbodal"       "Storgoteborg"  

table(areas)

#> areas
#>       Fyrbodal        Kungalv  Norra_Halland      Skaraborg Sodra_Alvsborg 
#>              9              6              3             12              9 
#>   Storgoteborg 
#>              7 

# But if we want Kungalv to be a part of Storgoteborg
# (which is common practice for example with lung cancer data):
(areas2 <- decode(healtcare_areas_west, "sjukvardsomrade", "kungalv2Storgoteborg"))

#> Warning: transformed to match the keyvalue:  Punctuations are removed. Only the first 4 characters are used.

#> Warning: Some codes could not be translated (54 cells)

#>   [1] NA               NA               NA               NA              
#>   [5] NA               NA               "Fyrbodal"       NA              
#>   [9] "Sodra_Alvsborg" "Skaraborg"      "Storgoteborg"   NA              
#>  [13] NA               "Sodra_Alvsborg" "Skaraborg"      "Sodra_Alvsborg"
#>  [17] "Storgoteborg"   "Storgoteborg"   NA               NA              
#>  [21] "Storgoteborg"   NA               NA               "Storgoteborg"  
#>  [25] "Storgoteborg"   NA               "Sodra_Alvsborg" "Skaraborg"     
#>  [29] NA               NA               "Fyrbodal"       NA              
#>  [33] "Storgoteborg"   NA               NA               "Skaraborg"     
#>  [37] NA               "Skaraborg"      "Norra_Halland"  "Skaraborg"     
#>  [41] NA               NA               NA               NA              
#>  [45] "Storgoteborg"   NA               NA               "Skaraborg"     
#>  [49] "Fyrbodal"       "Norra_Halland"  NA               NA              
#>  [53] NA               "Sodra_Alvsborg" "Fyrbodal"       "Storgoteborg"  
#>  [57] "Fyrbodal"       "Skaraborg"      "Skaraborg"      NA              
#>  [61] NA               NA               "Storgoteborg"   "Storgoteborg"  
#>  [65] "Sodra_Alvsborg" "Sodra_Alvsborg" NA               NA              
#>  [69] NA               NA               NA               NA              
#>  [73] "Skaraborg"      NA               NA               "Sodra_Alvsborg"
#>  [77] "Norra_Halland"  NA               NA               NA              
#>  [81] NA               NA               "Skaraborg"      "Storgoteborg"  
#>  [85] NA               NA               NA               NA              
#>  [89] NA               NA               "Fyrbodal"       NA              
#>  [93] "Sodra_Alvsborg" NA               "Fyrbodal"       "Fyrbodal"      
#>  [97] NA               "Skaraborg"      "Fyrbodal"       "Storgoteborg"  

table(areas2)

#> areas2
#>       Fyrbodal  Norra_Halland      Skaraborg Sodra_Alvsborg   Storgoteborg 
#>              9              3             12              9             13 

# We can also combine several extra_functions if we for example
# also want the area names with correct Swedish spelling.
(areas3 <- decode(healtcare_areas_west, "sjukvardsomrade", c("kungalv2Storgoteborg", "real_names")))

#> Warning: transformed to match the keyvalue:  Punctuations are removed. Only the first 4 characters are used.

#> Warning: Some codes could not be translated (54 cells)

#>   [1] NA               NA               NA               NA              
#>   [5] NA               NA               "Fyrbodal"       NA              
#>   [9] "Södra Älvsborg" "Skaraborg"      "Storgöteborg"   NA              
#>  [13] NA               "Södra Älvsborg" "Skaraborg"      "Södra Älvsborg"
#>  [17] "Storgöteborg"   "Storgöteborg"   NA               NA              
#>  [21] "Storgöteborg"   NA               NA               "Storgöteborg"  
#>  [25] "Storgöteborg"   NA               "Södra Älvsborg" "Skaraborg"     
#>  [29] NA               NA               "Fyrbodal"       NA              
#>  [33] "Storgöteborg"   NA               NA               "Skaraborg"     
#>  [37] NA               "Skaraborg"      "Norra Halland"  "Skaraborg"     
#>  [41] NA               NA               NA               NA              
#>  [45] "Storgöteborg"   NA               NA               "Skaraborg"     
#>  [49] "Fyrbodal"       "Norra Halland"  NA               NA              
#>  [53] NA               "Södra Älvsborg" "Fyrbodal"       "Storgöteborg"  
#>  [57] "Fyrbodal"       "Skaraborg"      "Skaraborg"      NA              
#>  [61] NA               NA               "Storgöteborg"   "Storgöteborg"  
#>  [65] "Södra Älvsborg" "Södra Älvsborg" NA               NA              
#>  [69] NA               NA               NA               NA              
#>  [73] "Skaraborg"      NA               NA               "Södra Älvsborg"
#>  [77] "Norra Halland"  NA               NA               NA              
#>  [81] NA               NA               "Skaraborg"      "Storgöteborg"  
#>  [85] NA               NA               NA               NA              
#>  [89] NA               NA               "Fyrbodal"       NA              
#>  [93] "Södra Älvsborg" NA               "Fyrbodal"       "Fyrbodal"      
#>  [97] NA               "Skaraborg"      "Fyrbodal"       "Storgöteborg"  

# The region names can be both with and without prefix:
regs <- sample(6, 10, replace = TRUE)
decode(regs, "region") # With prefix

#>  [1] "Region Syd"            "Region Väst"           "Region Syd"           
#>  [4] "Region Sydöstra"       "Region Norr"           "Region Uppsala/Örebro"
#>  [7] "Region Uppsala/Örebro" "Region Uppsala/Örebro" "Region Väst"          
#> [10] "Region Uppsala/Örebro"

decode(regs, "region", "short_region_names") # without prefix

#>  [1] "Syd"            "Väst"           "Syd"            "Sydöstra"      
#>  [5] "Norr"           "Uppsala/Örebro" "Uppsala/Örebro" "Uppsala/Örebro"
#>  [9] "Väst"           "Uppsala/Örebro"

# Note that only the first four digits of the LKF-code were used abowe?
# What if we use the full LKF-code?
lkfs <- sample(decoder:::forsamling$key, 100, replace = TRUE)
decode(lkfs, "sjukvardsomrade")

#> Warning: transformed to match the keyvalue:  Leading 0:s are ignored. Only the first 4 characters are used.

#> Warning: Some codes could not be translated (81 cells)

#>   [1] NA              NA              NA              NA             
#>   [5] NA              NA              NA              NA             
#>   [9] NA              NA              NA              NA             
#>  [13] "Storgoteborg"  NA              NA              NA             
#>  [17] NA              NA              NA              "Skaraborg"    
#>  [21] NA              NA              NA              NA             
#>  [25] NA              NA              NA              "Fyrbodal"     
#>  [29] NA              NA              NA              NA             
#>  [33] "Fyrbodal"      NA              NA              NA             
#>  [37] NA              NA              NA              NA             
#>  [41] NA              NA              NA              "Fyrbodal"     
#>  [45] "Skaraborg"     "Skaraborg"     NA              NA             
#>  [49] "Fyrbodal"      NA              NA              NA             
#>  [53] NA              NA              NA              NA             
#>  [57] NA              NA              NA              NA             
#>  [61] NA              NA              NA              NA             
#>  [65] "Norra_Halland" NA              "Skaraborg"     NA             
#>  [69] NA              NA              NA              NA             
#>  [73] "Fyrbodal"      NA              NA              NA             
#>  [77] NA              NA              NA              NA             
#>  [81] "Kungalv"       NA              NA              NA             
#>  [85] NA              NA              NA              NA             
#>  [89] "Fyrbodal"      "Skaraborg"     "Fyrbodal"      NA             
#>  [93] "Norra_Halland" "Storgoteborg"  NA              NA             
#>  [97] "Storgoteborg"  "Fyrbodal"      NA              NA             

# That work's just as fine when argument exact = FALSE (which it is by default).

# decode can also be used for data.frames with recognised column names
d <- data.frame(
     kon = sample(1:2, 10, replace = TRUE),
     sex = sample(1:2, 10, replace = TRUE),
     lkf = sample(decoder:::hemort$key, 10, replace = TRUE)
 )
 decode(d)

#> Warning: lkf -> lkf_kommun_beskrivning: transformed to match the keyvalue:  Only the first 4 characters are used.

#> Warning: lkf -> lkf_forsamling_beskrivning: Some codes could not be translated (7 cells)

#> Warning: lkf -> lkf_hemort2_beskrivning: Some codes could not be translated (7 cells)

#> Warning: lkf -> lkf_lan_beskrivning: transformed to match the keyvalue:  Only the first 2 characters are used.

#> New decoded columns added: 
#> * kon_kon_beskrivning
#> * lkf_kommun_beskrivning
#> * lkf_forsamling_beskrivning
#> * lkf_hemort2_beskrivning
#> * lkf_lan_beskrivning
#> * lkf_hemort_beskrivning
#> * sex_kon_beskrivning

#>    kon sex    lkf kon_kon_beskrivning lkf_kommun_beskrivning
#> 1    1   2 076405                 Man                Alvesta
#> 2    2   1 018016              Kvinna              Stockholm
#> 3    2   2 152909              Kvinna                   <NA>
#> 4    2   2 128618              Kvinna                  Ystad
#> 5    2   2 156205              Kvinna                   <NA>
#> 6    2   2 176505              Kvinna                 Årjäng
#> 7    1   1 030306                 Man                   <NA>
#> 8    2   1 147017              Kvinna                   Vara
#> 9    1   1 149120                 Man             Ulricehamn
#> 10   1   1 221301                 Man                   <NA>
#>    lkf_forsamling_beskrivning lkf_hemort2_beskrivning  lkf_lan_beskrivning
#> 1                      Moheda                  Moheda       Kronobergs län
#> 2                        <NA>                    <NA>       Stockholms län
#> 3                        <NA>                    <NA>                 <NA>
#> 4                        <NA>                    <NA>            Skåne län
#> 5                        <NA>                    <NA>                 <NA>
#> 6                    Blomskog                Blomskog        Värmlands län
#> 7                        <NA>                    <NA>          Uppsala län
#> 8                        <NA>                    <NA> Västra Götalands län
#> 9                      Hössna                  Hössna Västra Götalands län
#> 10                       <NA>                    <NA>  Västernorrlands län
#>    lkf_hemort_beskrivning sex_kon_beskrivning
#> 1                  Moheda              Kvinna
#> 2                 Högalid                 Man
#> 3                  Vesene              Kvinna
#> 4              Valleberga              Kvinna
#> 5              Ale-skövde              Kvinna
#> 6                Blomskog              Kvinna
#> 7              Härkeberga                 Man
#> 8              Främmestad                 Man
#> 9                  Hössna                 Man
#> 10                 Högsjö                 Man

### --- code --- ###
# Sometimes we have keyvalue objects with some key-value pairs without a 1:1 relation.
# This is true for snomed
# Show all non 1:1 pairs:
summary(decoder:::snomed)$nonunique

#>        key                                         value
#> 393  90703                           Embryonalt carcinom
#> 402  90723                           Embryonalt carcinom
#> 592  96153                               Hårcellsleukemi
#> 787  99403                               Hårcellsleukemi
#> 354  89643                               Klarcellssarkom
#> 381  90443                               Klarcellssarkom
#> 587  96073                             Mantelcellslymfom
#> 646  96733                             Mantelcellslymfom
#> 710  97423                              Mastcellsleukemi
#> 769  99003                              Mastcellsleukemi
#> 209  85103                               Medullär cancer
#> 215  85113                               Medullär cancer
#> 392  90701                 Misstänkt embryonalt carcinom
#> 396  90721                 Misstänkt embryonalt carcinom
#> 353  89641                     Misstänkt klarcellssarkom
#> 378  90441                     Misstänkt klarcellssarkom
#> 707  97333                            Plasmacellsleukemi
#> 741  98303                            Plasmacellsleukemi
#> 162  84001                     Svettkörtelcancer in situ
#> 163  84002                     Svettkörtelcancer in situ
#> 168  84101                      Talgkörtelcancer in situ
#> 171  84102                      Talgkörtelcancer in situ
#> 73  813021 Urotelialt carcinom, papillärt, icke invasivt
#> 74  813022 Urotelialt carcinom, papillärt, icke invasivt
#> 75  813031      Urotelialt carcinom, papillärt, invasivt
#> 76  813032      Urotelialt carcinom, papillärt, invasivt
#> 77  813033      Urotelialt carcinom, papillärt, invasivt
#> 79  812031  Urotelialt carcinom, UNS, inklusive cytologi
#> 80  812032  Urotelialt carcinom, UNS, inklusive cytologi
#> 81  812033  Urotelialt carcinom, UNS, inklusive cytologi

# Save them for later:
non_unique_snomeds <- summary(decoder:::snomed)$nonunique$key

# Use these snomed codes for decoding and coding
# Decoding works fine (all keys are unique) ...
(a <- decode(non_unique_snomeds, "snomed"))

#>  [1] "Embryonalt carcinom"                          
#>  [2] "Embryonalt carcinom"                          
#>  [3] "Hårcellsleukemi"                              
#>  [4] "Hårcellsleukemi"                              
#>  [5] "Klarcellssarkom"                              
#>  [6] "Klarcellssarkom"                              
#>  [7] "Mantelcellslymfom"                            
#>  [8] "Mantelcellslymfom"                            
#>  [9] "Mastcellsleukemi"                             
#> [10] "Mastcellsleukemi"                             
#> [11] "Medullär cancer"                              
#> [12] "Medullär cancer"                              
#> [13] "Misstänkt embryonalt carcinom"                
#> [14] "Misstänkt embryonalt carcinom"                
#> [15] "Misstänkt klarcellssarkom"                    
#> [16] "Misstänkt klarcellssarkom"                    
#> [17] "Plasmacellsleukemi"                           
#> [18] "Plasmacellsleukemi"                           
#> [19] "Svettkörtelcancer in situ"                    
#> [20] "Svettkörtelcancer in situ"                    
#> [21] "Talgkörtelcancer in situ"                     
#> [22] "Talgkörtelcancer in situ"                     
#> [23] "Urotelialt carcinom, papillärt, icke invasivt"
#> [24] "Urotelialt carcinom, papillärt, icke invasivt"
#> [25] "Urotelialt carcinom, papillärt, invasivt"     
#> [26] "Urotelialt carcinom, papillärt, invasivt"     
#> [27] "Urotelialt carcinom, papillärt, invasivt"     
#> [28] "Urotelialt carcinom, UNS, inklusive cytologi" 
#> [29] "Urotelialt carcinom, UNS, inklusive cytologi" 
#> [30] "Urotelialt carcinom, UNS, inklusive cytologi" 

# ... but coding these values back to their key does not

# NOT RUN {
code(a, "snomed")
# }